1) Change static variable into class variable to hold driver initialization    state. 2) Add mutex to Log::Write output just in case. Only done in Unix platform    code. Windows may or may not need this but I did see lost output on    Unix. 3) Value.h would read from freed memory in the case the reference count hit    zero. Probably doesn't matter but better form. 4) After stopping a thread, yield execution so that thread can terminate.    Just in case measure in unix platform code only. Unclear if necessary    for windows either. 5) Bug fix in variable arguments in Unix Log handling that might leak memory. 6) Missing memory free/delete in Node.cpp. 7) Add a string to Thread and ThreadImpl class to describe thread name for    debugging. 8) The biggest challenge was getting the Driver deconstructor to free    all its data structures properly without referencing free memory. Don't    mess with the order without checking for bad memory references.
Also the lock in the Unix Log implementation has some bad interactions
with output buffering. This doesn't occur in the log output file since it is
opened and closed every time. However stdout would flush its buffer at odd
times which somehow kept the lock from being released. If the thread exited
the rest of the threads would hang waiting for the lost lock. Now stdout
is flushed after every output occurance. Not a great performance trade-off
but should be okay unless we overuse the log.



git-svn-id: http://open-zwave.googlecode.com/svn/trunk@91 db0adee0-878e-11de-bed6-6d316fbe0868