Fix for two Thermostat / XML issues.
1)  ThermostatMode modes weren't properly read from the zwcfg xml file.  This was addressed by changing ThermostatMode::CreateVars to process differently based on how (from where) it was called.  To support this, I added Node:GetCurrentQueryStage(), which may be useful in other contexts as well.  I'm not too happy about making CreateVars public, but didn't see another straightforward way.  Ideas?

2)  ThermostatSetpoint had the same problem and was more problematic since the setpoint array only has instances (and values) for supported setpoints.  Which setpoints are supported isn't known before ReadXML is called.  So I adjusted Node::ReadValueFromXML and added another CreateVars virtual function (this one with two arguments..._instance and _index) to allow ThermostatSetpoint to create the necessary values.  It isn't pretty, but it works (at least for me, so far).

If someone has a better / prettier way of accomplishing this, be my guest.  The test for success is whether the a zwcfg XML file created from scratch is essentially identical to the one written after an existing zwcfg file is read and then re-written.  (I just run MinOZW twice in a row...once with no zwcfg file in the directory and the second with the one written by the first MinOZW run left in place.)

git-svn-id: http://open-zwave.googlecode.com/svn/trunk@199 db0adee0-878e-11de-bed6-6d316fbe0868