Some classes' values are retrieved via the Index component of ValueID; others are retrieved via the Instance component.  So all classes that implement RequestValue have been changed to accept two parameters (index and instance).  Lots of file changes, but only one fundamental concept change.
I've also given a try to recombining the VS2008 and VS2010 source code for OZWForm.  With a simple batch file (and a readme.txt file) for copying the solution and project files for a user's compiler into the .\src directory.

Hopefully it'll work with VS2010...otherwise we'll have to revert to the earlier directory structure.

git-svn-id: http://open-zwave.googlecode.com/svn/trunk@227 db0adee0-878e-11de-bed6-6d316fbe0868