Makefile updates with new targets/support: Make install - install into (default) directories (/usr/local/lib[64]/, /usr/local/etc, /usr/local/share/doc/openzwave-[version] Make dist - make a tarball that we can use to distribute
Version Number is now set at top of Makefile, and svn revision is pulled in. (reworked vers.cpp creation a bit)
Version Number is exported from Manager Class

'BUILD=[release/debug] make' automatically sets what type of build you want. Defaults to release mode now

Autodetect OSX/Linux in Makefile (so we can just have a single Makefile for both platforms - Needs testing and MinOZW not migrated to this yet... only the library and from the cpp/build/linux directory currently...

Version number is set in the soname of the shared library

shared library is installed with .so and .so.[version] number 

Made the build progress a little bit less verbose 

(Windows Builds are broken currently with the vers.cpp change... that will be fixed in next commit)



git-svn-id: http://open-zwave.googlecode.com/svn/branches/2013-11-13_release_testing@706 db0adee0-878e-11de-bed6-6d316fbe0868